GAME_FIELD in component
