+++ client/src/pages/Game/Game.tsx
@@ -16,6 +16,7 @@ import PlayerGameControls from '../../components/PlayerGameControls/PlayerGameCo
 import LeaveRoomModal from '../../components/LeaveRoomModal/LeaveRoomModal';
 import BetModal from '../../components/BetModal/BetModal';
 import RoomCodeMessage from '../../components/RoomCodeMessage/RoomCodeMessage';
+import GameTable from '../../components/GameTable/GameTable';
 import useGetCardImage from './hooks/useGetCardImage';
 
 
@@ -295,64 +296,14 @@ const GamePage: React.FC<IBasePage> = (props: IBasePage) => {
             <img src={tableImgSrc} alt="Poker Table" className="game-table-image" />
         </div>
         
-        {/* надписи с инфой игроков за столом */}
         <div id={GAME_FIELD} className={GAME_FIELD}>
-            <div className="players">
-                {players.map((player, index) => {
-                    const positions = ['left-top', 'left-middle', 'left-bottom', 'right-bottom', 'right-middle', 'right-top'];
-                    const position = positions[index] || 'left-top';
-                    const isCurrentPlayer = myMemberId !== null && player.memberId === myMemberId;
-                    const isActiveTurn = currentPlayerId === player.memberId;
-                    const score = playersScores[player.memberId] ?? 0;
-                    
-                    return (
-                        <div 
-                            className={`player-slot ${position} ${isCurrentPlayer ? 'active-player' : ''} ${isActiveTurn ? 'current-turn' : ''}`} 
-                            key={player.memberId}
-                        >
-                            <span className="name">{player.name}</span>
-                            <span className="balance">${player.balance}</span>
-                            <span className="score">{score}</span>
-                            {player.bet > 0 && <span className="bet">Ставка: ${player.bet}</span>}
-                        </div>
-                    );
-                })}
-            </div>
-
-            {/* карты игроков на столе */}
-            <div className="table-cards-container">
-                {players.map((player, index) => {
-                    const positions = ['left-top', 'left-middle', 'left-bottom', 'right-bottom', 'right-middle', 'right-top'];
-                    const position = positions[index] || 'left-top';
-                    const playerCards = player.cards ? player.cards.match(/.{1,2}/g) : [];
-                    
-                    return (
-                        <div 
-                            className={`table-player-cards ${position}`} 
-                            key={`table-cards-${player.memberId}`}
-                        >
-                            {playerCards && playerCards.length > 0 ? (
-                                playerCards.map((card, cardIndex) => {
-                                    const cardImage = getCardImage(card);
-                                    return (
-                                        <div 
-                                            className="table-card" 
-                                            key={cardIndex}
-                                            style={{ animationDelay: `${cardIndex * 0.2}s` }}
-                                        >
-                                            {cardImage ? (
-                                                <img src={cardImage} alt={card} />
-                                            ) : (
-                                                <span>{card}</span>
-                                            )}
-                                        </div>
-                                    );
-                                })
-                            ) : null}
-                        </div>
-                    );
-                })}
-            </div>
+            <GameTable
+                players={players}
+                playersScores={playersScores}
+                myMemberId={myMemberId}
+                currentPlayerId={currentPlayerId}
+                getCardImage={getCardImage}
+            />
              {/* дилер */}
              <div className='diller-slot'>
                 <span className="diller-name">Дилер: </span>
